more creature renames
