+++ assets/units/shouts/Kraken.ogg
