+++ src/index.html
@@ -442,9 +442,9 @@
 				<!-- TODO: This horrible mess needs to be coded procedually using JavaScript -->
 				<div id="creaturegridwrapper">
 					<div id="creaturegrid">
-						<a href="#Swine_Thug" class="vignette realmA typeA1" creature="A1">
+						<a href="#Thug_Swine" class="vignette realmA typeA1" creature="A1">
 							<div class="tooltip">
-								<div class="content">Swine Thug</div>
+								<div class="content">Thug Swine</div>
 							</div>
 							<div class="overlay"></div>
 							<div class="border"></div>
@@ -491,16 +491,16 @@
 							<div class="overlay"></div>
 							<div class="border"></div>
 						</a>
-						<a href="#Cycloper" class="vignette realmE typeE1" creature="E1">
+						<a href="#Mr._Stitches" class="vignette realmE typeE1" creature="E1">
 							<div class="tooltip">
-								<div class="content">Cycloper</div>
+								<div class="content">Mr. Stitches</div>
 							</div>
 							<div class="overlay"></div>
 							<div class="border"></div>
 						</a>
-						<a href="#Mr._Stitches" class="vignette realmE typeE2" creature="E2">
+						<a href="#Cycloper" class="vignette realmE typeE2" creature="E2">
 							<div class="tooltip">
-								<div class="content">Mr. Stitches</div>
+								<div class="content">Cycloper</div>
 							</div>
 							<div class="overlay"></div>
 							<div class="border"></div>
@@ -694,7 +694,7 @@
 							<div class="overlay"></div>
 							<div class="border"></div>
 						</a>
-						<a href="#Blue Shrimp" class="vignette realmS typeS2" creature="S2">
+						<a href="#Blue_Shrimp" class="vignette realmS typeS2" creature="S2">
 							<div class="tooltip">
 								<div class="content">Blue Shrimp</div>
 							</div>
@@ -722,17 +722,17 @@
 							<div class="overlay"></div>
 							<div class="border"></div>
 						</a>
-						<a href="#Squid Face" class="vignette
+						<a href="#Kraken" class="vignette
 								realmS typeS6" creature="S6">
 							<div class="tooltip">
-								<div class="content">Squid Face</div>
+								<div class="content">Kraken</div>
 							</div>
 							<div class="overlay"></div>
 							<div class="border"></div>
 						</a>
-						<a href="#Ice_Demon" class="vignette realmS typeS7" creature="S7">
+						<a href="#Vehemoth" class="vignette realmS typeS7" creature="S7">
 							<div class="tooltip">
-								<div class="content">Ice Demon</div>
+								<div class="content">Vehemoth</div>
 							</div>
 							<div class="overlay"></div>
 							<div class="border"></div>
@@ -792,8 +792,8 @@
 			</div>
 
 			<div id="musicplayerwrapper">
-				<!-- TODO: This list needs to be generated procedually, perhaps using Grunt -->
-				<audio id="audio" preload="auto" controls="" style="width: 890px;"></audio><a style="cursor:pointer;" id="mp_shuffle">Shuffle</a>
+				<!-- TODO: This list needs to be generated procedually -->
+				<audio id="audio" preload="auto" controls="" style="width: 890px;"></audio><a style="cursor: pointer;" id="mp_shuffle">Shuffle</a>
 				<ul id="playlist" style="list-style-type:none; padding-left:0px;">
 					<li class="active"><a href="<%= require('assets/music/Castle Black by Agret Brisignr.ogg') %>">Castle Black by Agret Brisignr</a></li>
 					<li class="active"><a href="<%= require('assets/music/City of Sand by Agret Brisignr.ogg') %>">City of Sand by Agret Brisignr</a></li>
