+++ src/style/avatars.less
@@ -1,6 +1,6 @@
 .vignette {
 	&.typeA1 {
-		background-image: url('~assets/units/avatars/Swine Thug.jpg');
+		background-image: url('~assets/units/avatars/Thug Swine.jpg');
 	}
 
 	&.typeA2 {
@@ -28,11 +28,11 @@
 	}
 
 	&.typeE1 {
-		background-image: url('~assets/units/avatars/Cycloper.jpg');
+		background-image: url('~assets/units/avatars/Mr. Stitches.jpg');
 	}
 
 	&.typeE2 {
-		background-image: url('~assets/units/avatars/Mr. Stitches.jpg');
+		background-image: url('~assets/units/avatars/Cycloper.jpg');
 	}
 
 	&.typeE3 {
@@ -160,11 +160,11 @@
 	}
 
 	&.typeS6 {
-		background-image: url('~assets/units/avatars/Squid Face.jpg');
+		background-image: url('~assets/units/avatars/Kraken.jpg');
 	}
 
 	&.typeS7 {
-		background-image: url('~assets/units/avatars/Ice Demon.jpg');
+		background-image: url('~assets/units/avatars/Vehemoth.jpg');
 	}
 
 	&.typeW1 {
